pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-sns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno==1.4.6",
    "browser-use>=0.1.45",
    "duckduckgo-search>=8.0.1",
    "langchain-openai>=0.3.11",
    "openai>=1.78.0",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "playwright>=1.52.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
    pythonpath = ["src"]

[project.scripts]
auto-sns = "auto_sns_agent.main:main"
